test(checker): split direct alias global tests
