+++ crates/tsz-checker/src/state/type_analysis/cross_file_direct_alias_chain_globals_tests.rs
@@ -1,0 +1,223 @@
+use super::cross_file_direct_alias_chain_tests::{
+    parse_bound_source, with_two_file_state_with_libs,
+};
+use crate::context::{CheckerContext, CheckerOptions};
+use crate::state::CheckerState;
+use std::sync::Arc;
+use tsz_binder::symbol_flags;
+use tsz_solver::TypeId;
+
+#[test]
+fn direct_source_file_type_alias_lowers_unshadowed_global_function_reference() {
+    with_two_file_state_with_libs(
+        "export type FunctionKeys<T> = { [K in keyof T]-?: T[K] extends Function ? K : never }[keyof T];",
+        "import { FunctionKeys } from './target';",
+        &["es5.d.ts"],
+        |state, target_binder| {
+            let function_keys_sym = target_binder
+                .file_locals
+                .get("FunctionKeys")
+                .expect("FunctionKeys");
+            let (ty, params) = state
+                .direct_source_file_type_alias_result(function_keys_sym, Some(1), true)
+                .expect("unshadowed global Function references should lower directly");
+            assert_ne!(ty, TypeId::UNKNOWN);
+            assert_ne!(ty, TypeId::ERROR);
+            assert_eq!(params.len(), 1, "FunctionKeys should expose T");
+        },
+    );
+}
+
+#[test]
+fn direct_source_file_type_alias_lowers_unshadowed_global_generic_reference() {
+    with_two_file_state_with_libs(
+        "export type Keep<Obj, Key extends keyof Obj> = Pick<Obj, Key>;",
+        "import { Keep } from './target';",
+        &["es5.d.ts"],
+        |state, target_binder| {
+            let keep_sym = target_binder.file_locals.get("Keep").expect("Keep");
+            let (ty, params) = state
+                .direct_source_file_type_alias_result(keep_sym, Some(1), true)
+                .expect("unshadowed global generic type references should lower directly");
+            assert_ne!(ty, TypeId::UNKNOWN);
+            assert_ne!(ty, TypeId::ERROR);
+            assert_eq!(params.len(), 2, "Keep should expose Obj and Key");
+        },
+    );
+}
+
+#[test]
+fn direct_source_file_type_alias_lowers_delegate_visible_global_generic_reference() {
+    let (target_arena, mut target_binder, types) =
+        parse_bound_source("export type Keep<Obj> = Required<Obj>;");
+    {
+        let target_binder = Arc::get_mut(&mut target_binder).expect("unique target binder");
+        let required_sym = target_binder
+            .symbols
+            .alloc(symbol_flags::TYPE_ALIAS, "Required".to_string());
+        target_binder
+            .file_locals
+            .set("Required".to_string(), required_sym);
+    }
+    let (requester_arena, requester_binder, _) =
+        parse_bound_source("import { Keep } from './target';");
+    let ctx = CheckerContext::new(
+        requester_arena.as_ref(),
+        requester_binder.as_ref(),
+        &types,
+        "requester.ts".to_string(),
+        CheckerOptions::default(),
+    );
+    let mut state = CheckerState { ctx };
+    state.ctx.set_all_arenas(Arc::new(vec![
+        Arc::clone(&requester_arena),
+        Arc::clone(&target_arena),
+    ]));
+    state.ctx.set_all_binders(Arc::new(vec![
+        Arc::clone(&requester_binder),
+        Arc::clone(&target_binder),
+    ]));
+
+    let keep_sym = target_binder.file_locals.get("Keep").expect("Keep");
+    let (ty, params) = state
+        .direct_source_file_type_alias_result(keep_sym, Some(1), true)
+        .expect("delegate-visible global generic type references should lower directly");
+    assert_ne!(ty, TypeId::UNKNOWN);
+    assert_ne!(ty, TypeId::ERROR);
+    assert_eq!(params.len(), 1, "Keep should expose Obj");
+}
+
+#[test]
+fn direct_source_file_type_alias_lowers_global_generic_reference_with_namespace_shadow() {
+    with_two_file_state_with_libs(
+        "namespace Pick {}\nexport type Keep<Obj, Key extends keyof Obj> = Pick<Obj, Key>;",
+        "import { Keep } from './target';",
+        &["es5.d.ts"],
+        |state, target_binder| {
+            let keep_sym = target_binder.file_locals.get("Keep").expect("Keep");
+            let (ty, params) = state
+                .direct_source_file_type_alias_result(keep_sym, Some(1), true)
+                .expect("namespace-only locals should not shadow global type aliases");
+            assert_ne!(ty, TypeId::UNKNOWN);
+            assert_ne!(ty, TypeId::ERROR);
+            assert_eq!(params.len(), 2, "Keep should expose Obj and Key");
+        },
+    );
+}
+
+#[test]
+fn direct_source_file_type_alias_lowers_global_generic_reference_with_value_shadow() {
+    with_two_file_state_with_libs(
+        "const Pick = 1;\nexport type Keep<Obj, Key extends keyof Obj> = Pick<Obj, Key>;",
+        "import { Keep } from './target';",
+        &["es5.d.ts"],
+        |state, target_binder| {
+            let keep_sym = target_binder.file_locals.get("Keep").expect("Keep");
+            let (ty, params) = state
+                .direct_source_file_type_alias_result(keep_sym, Some(1), true)
+                .expect("value-only locals should not shadow global type aliases");
+            assert_ne!(ty, TypeId::UNKNOWN);
+            assert_ne!(ty, TypeId::ERROR);
+            assert_eq!(params.len(), 2, "Keep should expose Obj and Key");
+        },
+    );
+}
+
+#[test]
+fn direct_source_file_type_alias_lowers_pick_by_value_shape_with_namespace_shadow() {
+    with_two_file_state_with_libs(
+        "import { Primitive } from './aliases-and-guards';\nnamespace Pick {}\nexport type PickByValue<T, ValueType> = Pick<T, { [Key in keyof T]-?: T[Key] extends ValueType ? Key : never }[keyof T]>;",
+        "import { PickByValue } from './target';",
+        &["es5.d.ts"],
+        |state, target_binder| {
+            let pick_by_value_sym = target_binder
+                .file_locals
+                .get("PickByValue")
+                .expect("PickByValue");
+            let (ty, params) = state
+                .direct_source_file_type_alias_result(pick_by_value_sym, Some(1), true)
+                .expect("utility-style PickByValue aliases should lower directly");
+            assert_ne!(ty, TypeId::UNKNOWN);
+            assert_ne!(ty, TypeId::ERROR);
+            assert_eq!(params.len(), 2, "PickByValue should expose T and ValueType");
+        },
+    );
+}
+
+#[test]
+fn direct_source_file_type_alias_rejects_local_type_alias_namespace_merge_shadow() {
+    with_two_file_state_with_libs(
+        "namespace Pick {}\ntype Pick<T, K> = T;\nexport type Keep<Obj, Key extends keyof Obj> = Pick<Obj, Key>;",
+        "import { Keep } from './target';",
+        &["es5.d.ts"],
+        |state, target_binder| {
+            let keep_sym = target_binder.file_locals.get("Keep").expect("Keep");
+            assert!(
+                state
+                    .direct_source_file_type_alias_result(keep_sym, Some(1), true)
+                    .is_none(),
+                "local type declarations merged with namespaces must not fall through to globals",
+            );
+        },
+    );
+}
+
+#[test]
+fn direct_source_file_type_alias_lowers_local_conditional_alias_argument_chain() {
+    with_two_file_state_with_libs(
+        "type SetDifference<A, B> = A extends B ? never : A;\ntype SetComplement<A, A1 extends A> = SetDifference<A, A1>;\nexport type FlowDiff<T extends U, U extends object> = Pick<T, SetComplement<keyof T, keyof U>>;",
+        "import { FlowDiff } from './target';",
+        &["es5.d.ts"],
+        |state, target_binder| {
+            let flow_diff_sym = target_binder.file_locals.get("FlowDiff").expect("FlowDiff");
+            let (ty, params) = state
+                .direct_source_file_type_alias_result(flow_diff_sym, Some(1), true)
+                .expect("local conditional alias argument chains should lower directly");
+            assert_ne!(ty, TypeId::UNKNOWN);
+            assert_ne!(ty, TypeId::ERROR);
+            assert_eq!(params.len(), 2, "FlowDiff should expose T and U");
+        },
+    );
+}
+
+#[test]
+fn direct_source_file_type_alias_lowers_intersection_of_local_and_global_applications() {
+    with_two_file_state_with_libs(
+        "type SetDifference<A, B> = A extends B ? never : A;\ntype Omit<T, K extends keyof any> = Pick<T, SetDifference<keyof T, K>>;\nexport type AugmentedRequired<T extends object, K extends keyof T = keyof T> = Omit<T, K> & Required<Pick<T, K>>;",
+        "import { AugmentedRequired } from './target';",
+        &["es5.d.ts"],
+        |state, target_binder| {
+            let augmented_required_sym = target_binder
+                .file_locals
+                .get("AugmentedRequired")
+                .expect("AugmentedRequired");
+            let (ty, params) = state
+                .direct_source_file_type_alias_result(augmented_required_sym, Some(1), true)
+                .expect("intersections of lowerable local and global generic applications should lower directly");
+            assert_ne!(ty, TypeId::UNKNOWN);
+            assert_ne!(ty, TypeId::ERROR);
+            assert_eq!(params.len(), 2, "AugmentedRequired should expose T and K");
+        },
+    );
+}
+
+#[test]
+fn direct_source_file_type_alias_lowers_utility_augmented_required_context() {
+    with_two_file_state_with_libs(
+        "import { Primitive } from './aliases-and-guards';\ntype SetDifference<A, B> = A extends B ? never : A;\nexport type Omit<T, K extends keyof any> = Pick<T, SetDifference<keyof T, K>>;\nexport type AugmentedRequired<T extends object, K extends keyof T = keyof T> = Omit<T, K> & Required<Pick<T, K>>;",
+        "import { AugmentedRequired } from './target';",
+        &["es5.d.ts"],
+        |state, target_binder| {
+            let augmented_required_sym = target_binder
+                .file_locals
+                .get("AugmentedRequired")
+                .expect("AugmentedRequired");
+            let (ty, params) = state
+                .direct_source_file_type_alias_result(augmented_required_sym, Some(1), true)
+                .expect("utility mapped-type context should lower directly");
+            assert_ne!(ty, TypeId::UNKNOWN);
+            assert_ne!(ty, TypeId::ERROR);
+            assert_eq!(params.len(), 2, "AugmentedRequired should expose T and K");
+        },
+    );
+}
